heavy refactor of template<->items logic - introduce mappings
Introduce a third primitive alongside templates and data: mappings.
Move the template selection to outside of items, and into a mappings
construct. The content/items specify the data. The templates specify
the, well, templates. The mappings represent a munging of template+data.
This enables the mapping to be treated separately, and multiple
templates can be applied to the same item data. Theoretically, output
arrangement also could have more flexibility.