snapshot of refactor to have a single mapping type
subitems for a mapping type (e.g. for indexes) should be a data
specification, not a mapping specification. Implement a way to define
subitems for an item. There's probably a bug in this, since we're
copying items by value instead of by reference...

Also item types should be a hardcoded field in the item keys. Instead,
items should be tagged based on arbitrary fields, and queries based on
arbitrary fields.

Implement a very poor selector expression language to support both
features.

Also rewrite the example project to be slightly more representative of a
real project.