specify full output path explicitly in mapping
Previously, and output file path's file extension was derived from the
extension on the template file. Change this to specify the full path
in the mapping file. For single-output rules, just specify the full path
explicitly. For multi-output rules, add the extension as part of the
JQ expression that we already use for generating the path names.